--- a/.github/workflows/ci-cd.yml
+++ b/.github/workflows/ci-cd.yml
@@ -22,7 +22,7 @@ jobs:
     needs: test
     if: github.event_name == 'push'
     runs-on: ubuntu-latest
-    environment: dev-prod # environment name of the repo
+    environment: dev-prod
 
     steps:
     # Checkout the code
@@ -38,9 +38,38 @@ jobs:
         node-version: '20.x'
     - run: npm ci
 
-    # # Install the dependencies
-    # - name: Install dependencies
-    #   run: npm install
+    # Configure AWS Credentials first
+    - name: Configure AWS credentials
+      uses: aws-actions/configure-aws-credentials@v1
+      with:
+        aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
+        aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+        aws-region: us-east-1
+
+    # Then login to ECR
+    - name: Login to Amazon ECR
+      id: login-ecr
+      uses: aws-actions/amazon-ecr-login@v1
+
+    # Build and push Docker images
+    - name: Build and push ECS images
+      env:
+        ECR_REGISTRY: ${{ steps.login-ecr.outputs.registry }}
+        AWS_REGION: us-east-1
+        SERVICE_NAME: prismix-serverless
+      run: |
+        # Set stage based on branch
+        STAGE=${GITHUB_REF##*/}
+        
+        # Build and push upload processor
+        docker build -t $ECR_REGISTRY/$SERVICE_NAME-upload-processor:$STAGE \
+          -f src/containers/upload-processor/Dockerfile .
+        docker push $ECR_REGISTRY/$SERVICE_NAME-upload-processor:$STAGE
+        
+        # Build and push archive processor
+        docker build -t $ECR_REGISTRY/$SERVICE_NAME-archive-processor:$STAGE \
+          -f src/containers/archive-processor/Dockerfile .
+        docker push $ECR_REGISTRY/$SERVICE_NAME-archive-processor:$STAGE
 
     # Set DATABASE_URL based on branch update
     - name: Set DATABASE_URL based on branch