--- a/.github/workflows/ci-cd.yml
+++ b/.github/workflows/ci-cd.yml
@@ -46,6 +46,29 @@ jobs:
         aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
         aws-region: us-east-1
 
+    # Create ECR repositories if they don't exist
+    - name: Create ECR repositories
+      env:
+        SERVICE_NAME: prismix-serverless
+      run: |
+        # Function to create repository if it doesn't exist
+        create_repo() {
+          local repo_name=$1
+          if ! aws ecr describe-repositories --repository-names $repo_name 2>/dev/null; then
+            echo "Creating repository $repo_name"
+            aws ecr create-repository \
+              --repository-name $repo_name \
+              --image-scanning-configuration scanOnPush=true \
+              --image-tag-mutability MUTABLE
+          else
+            echo "Repository $repo_name already exists"
+          fi
+        }
+        
+        # Create repositories
+        create_repo "$SERVICE_NAME-upload-processor"
+        create_repo "$SERVICE_NAME-archive-processor"
+
     # Then login to ECR
     - name: Login to Amazon ECR
       id: login-ecr