--- a/src/containers/archive-processor/Dockerfile
+++ b/src/containers/archive-processor/Dockerfile
@@ -1,6 +1,9 @@
-FROM node:18-alpine
+# Build stage
+FROM node:18-alpine AS builder
 
-# Install system dependencies
+WORKDIR /app
+
+# Install build dependencies
 RUN apk add --no-cache \
     zip \
     unzip \
@@ -8,25 +11,37 @@ RUN apk add --no-cache \
     py3-pip \
     build-base
 
-WORKDIR /app
-
-# Copy package files
+# Copy only package files first
 COPY package*.json ./
 
-# Install dependencies that were previously in layers
-RUN npm install \
+# Install only production dependencies
+RUN npm ci --only=production \
+    @aws-sdk/client-s3 \
+    @aws-sdk/client-dynamodb \
+    @aws-sdk/client-sqs \
+    @aws-sdk/lib-dynamodb \
     @supabase/supabase-js \
     archiver \
-    p-limit
+    p-limit \
+    sqs-consumer
 
-# Copy application code
+# Copy only necessary files
 COPY src/handlers/zipArchiveProcessor.js ./src/handlers/
 COPY src/services ./src/services
 COPY src/utils ./src/utils
-
-# Copy container-specific files
 COPY src/containers/archive-processor/index.js ./
-COPY src/containers/archive-processor/healthcheck.js ./
+COPY src/containers/healthcheck.js ./
+
+# Production stage
+FROM node:18-alpine
+
+# Install runtime dependencies
+RUN apk add --no-cache zip unzip
+
+WORKDIR /app
+
+# Copy from builder stage
+COPY --from=builder /app ./
 
 # Set up health check
 HEALTHCHECK --interval=30s --timeout=5s --start-period=60s --retries=3 \