--- a/src/containers/archive-processor/healthcheck.js
+++ b/src/containers/archive-processor/healthcheck.js
@@ -1,0 +1,10 @@
+const http = require('http');
+
+const server = http.createServer((req, res) => {
+    res.writeHead(200);
+    res.end('healthy');
+});
+
+server.listen(80, () => {
+    console.log('Health check server running');
+});