--- a/src/containers/archive-processor/package.json
+++ b/src/containers/archive-processor/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "archive-processor",
+  "version": "1.0.0",
+  "description": "ECS container for processing archive operations",
+  "main": "index.js",
+  "scripts": {
+    "start": "node index.js",
+    "health": "node healthcheck.js"
+  },
+  "dependencies": {
+    "@aws-sdk/client-s3": "^3.0.0",
+    "@aws-sdk/client-dynamodb": "^3.0.0",
+    "@aws-sdk/client-sqs": "^3.0.0",
+    "@aws-sdk/lib-dynamodb": "^3.0.0",
+    "@supabase/supabase-js": "^2.0.0",
+    "archiver": "^5.3.1",
+    "sqs-consumer": "^7.0.0",
+    "p-limit": "^3.1.0"
+  }
+}