--- a/src/containers/dependencies.json
+++ b/src/containers/dependencies.json
@@ -1,0 +1,12 @@
+{
+    "supabaseLayer2": {
+        "@supabase/supabase-js": "latest",
+        "other-supabase-dependencies": "version"
+    },
+    "archiver-layer": {
+        "archiver": "latest"
+    },
+    "p-limit-layer": {
+        "p-limit": "latest"
+    }
+}