--- a/src/containers/upload-processor/Dockerfile
+++ b/src/containers/upload-processor/Dockerfile
@@ -1,22 +1,33 @@
+# Build stage
+FROM node:18-alpine AS builder
+
+WORKDIR /app
+
+# Copy only package files first to leverage cache
+COPY package*.json ./
+
+# Install only production dependencies
+RUN npm ci --only=production \
+    @aws-sdk/client-s3 \
+    @aws-sdk/client-dynamodb \
+    @aws-sdk/lib-dynamodb \
+    @supabase/supabase-js \
+    sqs-consumer
+
+# Copy only necessary files
+COPY src/handlers/imageUploadFromS3Processor.js ./src/handlers/
+COPY src/services ./src/services
+COPY src/utils ./src/utils
+COPY src/containers/upload-processor/index.js ./
+COPY src/containers/healthcheck.js ./
+
+# Production stage
 FROM node:18-alpine
 
 WORKDIR /app
 
-# Copy package files
-COPY package*.json ./
-
-# Install dependencies that were previously in layers
-RUN npm install \
-    @supabase/supabase-js
-
-# Copy application code
-COPY src/handlers/imageUploadFromS3Processor.js ./src/handlers/
-COPY src/services ./src/services
-COPY src/utils ./src/utils
-
-# Copy container-specific files
-COPY src/containers/upload-processor/index.js ./
-COPY src/containers/upload-processor/healthcheck.js ./
+# Copy from builder stage
+COPY --from=builder /app ./
 
 # Set up health check
 HEALTHCHECK --interval=30s --timeout=5s --start-period=60s --retries=3 \