…rrency

- Increased memorySize for blurDetection Lambda from 2048 to 3008 for improved CPU allocation.
- Added reservedConcurrency of 200 and provisionedConcurrency of 10 to enhance capacity and response times.
- Updated CONCURRENT_QUALITY_CHECKS in qualityService.js from 50 to 100 to align with the new Lambda capacity.
- Reduced BASE_DELAY from 50ms to 25ms and MAX_BACKOFF from 1000ms to 500ms for faster request handling.
- Increased maxRequests in the request window from 100 to 200 to support higher throughput while maintaining rate control.